+++ daniel/src/main/java/app/domain/models/Transfer/Transfer.java
@@ -3,6 +3,8 @@ package app.domain.models.Transfer;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import app.domain.models.Transfer.enums.TransferStatus;
@@ -11,10 +13,10 @@ import app.domain.models.Transfer.enums.TransferStatus;
 @Getter
 @NoArgsConstructor
 public class Transfer { //transferencia
-    private long id; 
+    private long id;
     private String sourceAccount; //cuenta origen
     private String destinationAccount; //cuenta destino
-    private double amount; //monto
+    private BigDecimal amount; //monto
     private LocalDateTime creationDate; //fecha de creacion
     private LocalDateTime approvalDate; //fecha de aprobacion
     private TransferStatus transferStatus; //estado de la transferencia
